+++ cogs/starboard.py
@@ -1,9 +1,12 @@
 __author__ = 'Chris Taylor'
 __copyright__ = 'Copyright CEN'
-__credits__ = 'Chris Taylor'
-__version__ = '0.0.0'
-__status__ = 'Development'
+__credits__ = 'Chris Taylor, Justin Panchula'
+__version__ = '1.0.0'
+__status__ = 'Production'
 __doc__ = """Starboard functions"""
+
+# Python imports
+from datetime import datetime
 
 # Discord imports
 from cbot import cbot
@@ -23,7 +26,15 @@ class starboard(commands.GroupCog, name='starboard'):
     def __init__(self, bot: cbot) -> None:
         self.bot = bot
         super().__init__()
-    
+
+    def create_embed(self, message: discord.Message, awarder: str) -> discord.Embed:
+        embed = discord.Embed(color=discord.Color.gold(), title="Starboard Message")
+        embed.set_author(name=message.author.name, icon_url=message.author.display_avatar)
+        embed.add_field(name='Message', value=f"{message.content}\n[View Message]({message.jump_url})", inline=False)
+        embed.add_field(name="Awarded By", value=awarder, inline=False)
+        embed.set_footer(text=f"Awarded at: {datetime.now().strftime('%d/%m/%y - %H:%M:%S')}")
+        return embed
+
     # Set the starboard channel
     @app_commands.command(
         name='setchannel',
@@ -42,7 +53,7 @@ class starboard(commands.GroupCog, name='starboard'):
             await interaction.response.send_message("There was an error, please try again.", ephemeral=True)
         else:
             await interaction.response.send_message("Starboard channel set.", ephemeral=False)
-    
+
     # Set the starboard threshold
     @app_commands.command(
         name='setthreshold',
@@ -64,15 +75,59 @@ class starboard(commands.GroupCog, name='starboard'):
 
     # Add to starboard
     @commands.Cog.listener()
-    async def on_raw_reaction_add(payload: discord.RawReactionActionEvent):
+    async def on_raw_reaction_add(self, payload: discord.RawReactionActionEvent):
         # Parse payload
-        message = discord.utils. ()
-        guild = payload.guild_id
-        emoji = payload.emoji
+        guild = await self.bot.fetch_guild(payload.guild_id)
+        channel = await guild.fetch_channel(payload.channel_id)
+        message = await channel.fetch_message(payload.message_id)
+        member = await guild.fetch_member(payload.user_id)
+        emoji = payload.emoji.name
+
+        # Reject embed messages
+        if len(message.embeds) != 0:
+            return
+
+        # Get starboard channel
+        try:
+            async with self.bot.db_pool.acquire() as con:
+                response = await con.fetch("SELECT starboard_channel FROM serverdata WHERE guild_id=$1", guild.id)
+                channel = await self.bot.fetch_channel(response[0]['starboard_channel'])
+        except PostgresError as e:
+            logger.exception(e)
+            return
+        except AttributeError as e:
+            logger.exception(e)
+        except discord.DiscordException as e:
+            logger.exception(e)
+
+        # Get starboard threshold
+        try:
+            async with self.bot.db_pool.acquire() as con:
+                response = await con.fetch("SELECT starboard_threshold FROM serverdata WHERE guild_id=$1", guild.id)
+                threshold = response[0]['starboard_threshold']
+        except PostgresError as e:
+            logger.exception(e)
+            return
+        except AttributeError as e:
+            logger.exception(e)
+            return
 
         # Logic
-        if emoji == '⭐':
-            if message.
+        if emoji == '🌟' and member.guild_permissions.administrator:
+            embed = self.create_embed(message, member.name)
+            await channel.send(embed=embed)
+        elif emoji == '⭐':
+            stars = 0
+
+            # Get reactions
+            for reaction in message.reactions:
+                if reaction.emoji == '⭐':
+                    stars += 1
+
+            # Check and create
+            if stars == threshold:
+                embed = self.create_embed(message, f"{guild.name} Community")
+                await channel.send(embed=embed)
 
 
 # Add to bot
