+++ cogs_deprecated/create_table_react.sql
